fastapi_app_v5/server/src/requirements.txt
fastapi==0.98.0
gcsfs==2023.6.0
google_cloud_aiplatform==1.28.0
google-generativeai==0.1.0
langchain==0.0.237
openai==0.27.8
protobuf==4.23.3
python-dotenv==0.21.1
pyowm==3.3.0
uvicorn==0.22.0
